Project 3.0 - Exploiting Buffer Overflow Vulnerabilities/release/requirements.txt
fastapi[standard]
oracledb
pyjwt
passlib[bcrypt]
requests
pwntools